* severe optimization and profiling for DSP code. Now to unify unit tests
git-svn-id: file:///home/jonas/soma/svn-restore.20080710/DSPboard/trunk@2852 5dccc5d9-beb7-0310-bf71-cc18d937a3f4

+++ src/acqboardif.h
@@ -33,7 +33,7 @@ public:
 
   unsigned char mode; 
   bool linkUp; 
-  int gain[CHANNUM]; 
+  short gain[CHANNUM]; 
   bool hpfen[CHANNUM]; 
   char inputSel; 
 };
